refactor(tmux): move pane LC_CTYPE resolver into tmux::locale
Per PR review: relocate `utf8_ctype_for_panes` (and its pure helper
`resolve_utf8_ctype` plus their unit tests) from `src/main.rs` into
`src/tmux/locale.rs`, next to their sole caller `propagate_utf8_ctype`.
The function is now private since nothing outside the module uses it.

The shared primitives `is_utf8_locale` and `UTF8_CTYPE_FALLBACK` stay in
`src/main.rs` (still used by `ensure_utf8_locale_env` /
`utf8_locale_fallback` at startup) and are exposed `pub(crate)`. No
behavior change.
