+++ src/tmux/locale.rs
@@ -24,7 +24,7 @@
 //! untouched (overriding it would clobber a deliberate UTF-8 `LC_ALL` and other
 //! locale categories), so that rarer case is not repaired by this.
 
-use crate::utf8_ctype_for_panes;
+use crate::{UTF8_CTYPE_FALLBACK, is_utf8_locale};
 use bevy::prelude::*;
 use ozmux_tmux::{SetEnvironmentGlobal, TmuxClient, TmuxClientMut};
 
@@ -57,9 +57,74 @@ fn propagate_utf8_ctype(mut client: TmuxClientMut<'_, '_>) {
     }
 }
 
+/// The UTF-8 `LC_CTYPE` ozmux advertises to tmux panes spawned after attach, so
+/// their shells render typed wide characters correctly instead of as `<00xx>`
+/// placeholders (see this module's docs for the full failure mode).
+///
+/// Resolves to the first UTF-8 locale among `LC_ALL` / `LC_CTYPE` / `LANG`,
+/// skipping any non-UTF-8 entries, falling back to [`UTF8_CTYPE_FALLBACK`].
+///
+/// This deliberately differs from `utf8_locale_fallback` (in `src/main.rs`):
+/// that one honors tmux's first-non-empty-wins precedence to decide *whether*
+/// ozmux's own process needs the fallback, so a leading non-UTF-8 value (e.g.
+/// `LC_ALL=C`) makes it return the fallback. Here we instead want a concrete
+/// UTF-8 *value* to advertise to panes, so a non-UTF-8 entry is skipped rather
+/// than allowed to win. Because `ensure_utf8_locale_env` runs first, on macOS
+/// this always resolves to a UTF-8 value even when the inherited environment
+/// selected the C/POSIX locale.
+fn utf8_ctype_for_panes() -> String {
+    resolve_utf8_ctype(
+        std::env::var("LC_ALL").ok().as_deref(),
+        std::env::var("LC_CTYPE").ok().as_deref(),
+        std::env::var("LANG").ok().as_deref(),
+    )
+}
+
+/// Pure resolver behind [`utf8_ctype_for_panes`]: the first non-empty UTF-8
+/// locale among the three values, or [`UTF8_CTYPE_FALLBACK`].
+fn resolve_utf8_ctype(lc_all: Option<&str>, lc_ctype: Option<&str>, lang: Option<&str>) -> String {
+    [lc_all, lc_ctype, lang]
+        .into_iter()
+        .flatten()
+        .find(|value| !value.is_empty() && is_utf8_locale(value))
+        .map(str::to_string)
+        .unwrap_or_else(|| UTF8_CTYPE_FALLBACK.to_string())
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn resolve_ctype_falls_back_when_no_utf8_locale() {
+        assert_eq!(resolve_utf8_ctype(None, None, None), "en_US.UTF-8");
+        assert_eq!(
+            resolve_utf8_ctype(Some("C"), Some("C"), Some("C")),
+            "en_US.UTF-8"
+        );
+        assert_eq!(
+            resolve_utf8_ctype(Some(""), Some(""), Some("POSIX")),
+            "en_US.UTF-8"
+        );
+    }
+
+    #[test]
+    fn resolve_ctype_keeps_first_utf8_locale() {
+        // The user's own UTF-8 locale is preserved rather than overwritten with
+        // the en_US fallback, honoring tmux's LC_ALL > LC_CTYPE > LANG order.
+        assert_eq!(
+            resolve_utf8_ctype(None, Some("ja_JP.UTF-8"), None),
+            "ja_JP.UTF-8"
+        );
+        assert_eq!(
+            resolve_utf8_ctype(Some("en_US.UTF-8"), Some("C"), None),
+            "en_US.UTF-8"
+        );
+        assert_eq!(
+            resolve_utf8_ctype(Some("C"), None, Some("ja_JP.UTF-8")),
+            "ja_JP.UTF-8"
+        );
+    }
 
     #[test]
     fn propagate_sends_global_utf8_lc_ctype() {
